+++ features/operational/contact_us/src/test/java/com/tokopedia/contactus/inboxtickets/viewmodel/InboxDetailViewModelTest.kt
@@ -1526,7 +1526,7 @@ class InboxDetailViewModelTest {
         for (i in 1..6) {
             badCsatReasonListItem.add(
                 BadCsatReasonListItem().apply {
-                    id = i
+                    id = i.toLong()
                     messageEn = "message-$i"
                     message = "message-$i"
                 }
